requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
docopt==0.6.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
Js2Py==0.71
MarkupSafe==2.0.1
packaging==20.9
pipwin==0.5.1
psycopg2-binary==2.9.1
pyjsparser==2.7.1
PyJWT==2.1.0
pyparsing==2.4.7
PyPrind==2.11.3
pySmartDL==1.3.4
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.6
